Moving common fixtures to conftest
