+++ presidio-image-redactor/tests/integration/test_dicom_image_pii_verify_engine_integration.py
@@ -6,15 +6,12 @@ the original parent ImagePiiVerifyEngine class.
 """
 import PIL
 import pydicom
-import json
-import os
 
 from presidio_image_redactor.dicom_image_pii_verify_engine import (
     DicomImagePiiVerifyEngine,
 )
 import pytest
 
-SCRIPT_DIR = os.path.dirname(__file__)
 PADDING_WIDTH = 25
 
 
@@ -27,27 +24,6 @@ def mock_engine():
     dicom_image_pii_verify_engine = DicomImagePiiVerifyEngine()
 
     return dicom_image_pii_verify_engine
-
-
-@pytest.fixture(scope="module")
-def mock_instance():
-    """DICOM instance to use in testing"""
-    # Assign
-    filepath = f"{SCRIPT_DIR}/resources/0_ORIGINAL.dcm"
-
-    # Act
-    instance = pydicom.dcmread(filepath)
-
-    return instance
-
-
-@pytest.fixture(scope="module")
-def mock_results():
-    """Loaded json results file"""
-    with open(f"{SCRIPT_DIR}/resources/dicom_pii_verify_integration.json") as json_file:
-        results_json = json.load(json_file)
-
-    return results_json
 
 
 def test_verify_correctly(
